+++ i18n/fr.json
@@ -3,14 +3,16 @@
   "HOME": "Accueil",
   "LANGUAGE": "La langue",
   "ENGLISH": "Anglais",
-  "FRENCH": "Français",
-  "SEARCH": "Recherche",
-
+  "FRENCH": "Français",  
   "ORIGINGEOHASH": "Origine Geohash",
   "DESTINATIONGEOHASH": "Destination Geohash",
   "DEPARTUREDATE": "Date de départ",
   "WHENTOLEAVE": "Quand comptez-vous partir?",
-
+  "ADULTS": "Adultes",
+  "CHILDREN": "Enfants",
+  "SENIORS": "Aînés",
+  "SEARCH": "Recherche",
+  "SEARCHING": "Fouille...",
   "DEPARTURELOCATION": "Départ de",
   "DEPARTURETIME": "Départ à",
   "ARRIVALLOCATION": "Arriver à",
