+++ lib/busbud.js
@@ -1,6 +1,5 @@
 var axios = require("axios");
-var moment = require("moment-timezone");
-var accounting = require("accounting");
+var merge = require('deepmerge');
 
 var busbud = axios.create({
   baseURL: "https://napi.busbud.com/x-departures/",
@@ -11,52 +10,28 @@ var busbud = axios.create({
   }
 });
 
-function findLocation(locationId, locations) {
-  return locations.find(loc => loc.id === locationId);
-};
-
-function calculateTime(inTime, inTimeZone) {
-  return moment.tz(inTime, inTimeZone).format('LLLL');
-};
-
-function parseDeparture(departure, locations) {
-  let destinationLocation = findLocation(departure.destination_location_id, locations);
-  let originLocation = findLocation(departure.origin_location_id, locations);
-  let destinationTime = calculateTime(departure.arrival_time, departure.arrival_timezone);
-  let originTime = calculateTime(departure.departure_time, departure.departure_timezone);
-
-  return {
-    price: accounting.formatMoney(departure.prices.total / 100),
-    departTime: destinationTime,
-    arriveTime: originTime,
-    originLocationName: originLocation.name,
-    destinationLocationName: destinationLocation.name
-  };
-};
+function getRandomInt(min, max) {
+  min = Math.ceil(min);
+  max = Math.floor(max);
+  return Math.floor(Math.random() * (max - min)) + min;
+}
 
 function fetchData(source, destination, date, params, poll = false) {
-  let gatheredData = { locations: [], departures: [] };
+  let gatheredData = {};
   return new Promise(function(resolve, reject) {
     let requestUrl = "/" + source + "/" + destination + "/" + date + (poll ? "/poll" : "");
     busbud.get(requestUrl, params).then(function(response) {
-      gatheredData.locations = response.data.locations;
-      gatheredData.departures = response.data.departures;
-      
-      if (response.data.complete === false) {
+      gatheredData = response.data;
+      if (gatheredData.complete === false) {
         params.index = gatheredData.departures.length;
         setTimeout(() => {
           fetchData(source, destination, date, params, true).then(data => {
-            gatheredData.locations = gatheredData.locations.concat(data.locations);
-            gatheredData.departures = gatheredData.departures.concat(data.departures);
-
-            gatheredData.locations = Array.from(new Set(gatheredData.locations));
-            gatheredData.departures = Array.from(new Set(gatheredData.departures));
-
+            gatheredData = merge(gatheredData, data);
             resolve(gatheredData);
           }).catch(error => {
             reject(error);
           });
-        }, 2000);
+        }, getRandomInt(2000, 5000));
       } else {
         resolve(gatheredData);
       }
@@ -75,12 +50,9 @@ busbud.search = function(source, destination, date, options = {}) {
     currency: "CAD"
   };
 
-  params = Object.assign(params, options);
+  params = merge(params, options);
+  delete params.index;
   return fetchData(source, destination, date, params);
-};
-
-busbud.parseDepartures = function(results) {
-  return results.departures.map(departure => parseDeparture(departure, results.locations));
 };
 
 module.exports = busbud;
