+++ package.json
@@ -12,6 +12,7 @@
     "cookie-parser": "~1.4.3",
     "csurf": "^1.9.0",
     "debug": "~2.6.9",
+    "deepmerge": "^2.1.0",
     "ejs": "~2.5.7",
     "express": "~4.16.0",
     "express-session": "^1.15.6",
